requirements.txt
requests
python-dateutil
APScheduler
pytz
python-dotenv
beautifulsoup4
